+++ PyDodo/pydodo/__init__.py
@@ -1,10 +1,7 @@
 
-from .aircraft_position import aircraft_position
-from .all_positions import all_positions
-from .change_altitude import change_altitude
-from .change_heading import change_heading
-from .change_speed import change_speed
-from .change_vertical_speed import change_vertical_speed
+from .request_position import aircraft_position, all_positions
+from .request_change import (change_altitude, change_heading, change_speed,
+                             change_vertical_speed)
 from .create_aircraft import create_aircraft
 from .load_scenario import load_scenario
 from .reset_simulation import reset_simulation
